+++ dropshot_endpoint/Cargo.toml
@@ -24,4 +24,6 @@ version = "2.0"
 features = [ "full" ]
 
 [dev-dependencies]
+expectorate = "1.1.0"
+prettyplease = "0.2.20"
 schema = "0.1.0"
