+++ dropshot_endpoint/src/endpoint.rs
@@ -447,6 +447,9 @@ pub(crate) struct EndpointMetadata {
 
 #[cfg(test)]
 mod tests {
+    use expectorate::assert_contents;
+    use syn::parse_quote;
+
     use super::*;
 
     #[test]
@@ -465,82 +468,12 @@ mod tests {
             },
         )
         .unwrap();
-        let expected = quote! {
-            const _: fn() = || {
-                struct NeedRequestContext(<RequestContext<()> as dropshot::RequestContextArgument>::Context) ;
-            };
-            const _: fn() = || {
-                trait ResultTrait {
-                    type T;
-                    type E;
-                }
-                impl<TT, EE> ResultTrait for Result<TT, EE>
-                where
-                    TT: dropshot::HttpResponse,
-                {
-                    type T = TT;
-                    type E = EE;
-                }
-                struct NeedHttpResponse(
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::T,
-                );
-                trait TypeEq {
-                    type This: ?Sized;
-                }
-                impl<T: ?Sized> TypeEq for T {
-                    type This = Self;
-                }
-                fn validate_result_error_type<T>()
-                where
-                    T: ?Sized + TypeEq<This = dropshot::HttpError>,
-                {
-                }
-                validate_result_error_type::<
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::E,
-                >();
-            };
-
-            #[allow(non_camel_case_types, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            pub struct handler_xyz {}
-
-            #[allow(non_upper_case_globals, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            pub const handler_xyz: handler_xyz = handler_xyz {};
-
-            impl From<handler_xyz>
-                for dropshot::ApiEndpoint<
-                    <RequestContext<()>
-                as dropshot::RequestContextArgument>::Context>
-            {
-                fn from(_: handler_xyz) -> Self {
-                    #[allow(clippy::unused_async)]
-                    pub async fn handler_xyz(
-                        _rqctx: RequestContext<()>,
-                    ) -> Result<HttpResponseOk<()>, HttpError> {
-                        Ok(())
-                    }
-
-                    const _: fn() = || {
-                        fn future_endpoint_must_be_send<T: ::std::marker::Send>(_t: T) {}
-                        fn check_future_bounds(arg0: RequestContext<()>) {
-                            future_endpoint_must_be_send(handler_xyz(arg0));
-                        }
-                    };
-
-                    dropshot::ApiEndpoint::new(
-                        "handler_xyz".to_string(),
-                        handler_xyz,
-                        dropshot::Method::GET,
-                        "application/json",
-                        "/a/b/c",
-                    )
-                }
-            }
-        };
 
         assert!(errors.is_empty());
-        assert_eq!(expected.to_string(), item.to_string());
+        assert_contents(
+            "tests/output/endpoint_basic.rs",
+            &prettyplease::unparse(&parse_quote! { #item }),
+        );
     }
 
     #[test]
@@ -558,78 +491,12 @@ mod tests {
                 }
             },
         ).unwrap();
-        let expected = quote! {
-            const _: fn() = || {
-                struct NeedRequestContext(<dropshot::RequestContext<()> as dropshot::RequestContextArgument>::Context) ;
-            };
-            const _: fn() = || {
-                trait ResultTrait {
-                    type T;
-                    type E;
-                }
-                impl<TT, EE> ResultTrait for Result<TT, EE>
-                where
-                    TT: dropshot::HttpResponse,
-                {
-                    type T = TT;
-                    type E = EE;
-                }
-                struct NeedHttpResponse(
-                    <std::Result<dropshot::HttpResponseOk<()>, dropshot::HttpError> as ResultTrait>::T,
-                );
-                trait TypeEq {
-                    type This: ?Sized;
-                }
-                impl<T: ?Sized> TypeEq for T {
-                    type This = Self;
-                }
-                fn validate_result_error_type<T>()
-                where
-                    T: ?Sized + TypeEq<This = dropshot::HttpError>,
-                {
-                }
-                validate_result_error_type::<
-                    <std::Result<dropshot::HttpResponseOk<()>, dropshot::HttpError> as ResultTrait>::E,
-                >();
-            };
-
-            #[allow(non_camel_case_types, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            pub struct handler_xyz {}
-
-            #[allow(non_upper_case_globals, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            pub const handler_xyz: handler_xyz = handler_xyz {};
-
-            impl From<handler_xyz> for dropshot::ApiEndpoint< <dropshot::RequestContext<()> as dropshot::RequestContextArgument>::Context> {
-                fn from(_: handler_xyz) -> Self {
-                    #[allow(clippy::unused_async)]
-                    pub async fn handler_xyz(_rqctx: dropshot::RequestContext<()>) ->
-                        std::Result<dropshot::HttpResponseOk<()>, dropshot::HttpError>
-                    {
-                        Ok(())
-                    }
-
-                    const _: fn() = || {
-                        fn future_endpoint_must_be_send<T: ::std::marker::Send>(_t: T) {}
-                        fn check_future_bounds(arg0: dropshot::RequestContext<()>) {
-                            future_endpoint_must_be_send(handler_xyz(arg0));
-                        }
-                    };
-
-                    dropshot::ApiEndpoint::new(
-                        "handler_xyz".to_string(),
-                        handler_xyz,
-                        dropshot::Method::GET,
-                        "application/json",
-                        "/a/b/c",
-                    )
-                }
-            }
-        };
 
         assert!(errors.is_empty());
-        assert_eq!(expected.to_string(), item.to_string());
+        assert_contents(
+            "tests/output/endpoint_context_fully_qualified_names.rs",
+            &prettyplease::unparse(&parse_quote! { #item }),
+        );
     }
 
     #[test]
@@ -650,93 +517,12 @@ mod tests {
             },
         )
         .unwrap();
-        let expected = quote! {
-            const _: fn() = || {
-                struct NeedRequestContext(<RequestContext<std::i32> as dropshot::RequestContextArgument>::Context) ;
-            };
-            const _: fn() = || {
-                fn need_exclusive_extractor<T>()
-                where
-                    T: ?Sized + dropshot::ExclusiveExtractor,
-                {
-                }
-                need_exclusive_extractor::<Query<Q> >();
-            };
-            const _: fn() = || {
-                trait ResultTrait {
-                    type T;
-                    type E;
-                }
-                impl<TT, EE> ResultTrait for Result<TT, EE>
-                where
-                    TT: dropshot::HttpResponse,
-                {
-                    type T = TT;
-                    type E = EE;
-                }
-                struct NeedHttpResponse(
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::T,
-                );
-                trait TypeEq {
-                    type This: ?Sized;
-                }
-                impl<T: ?Sized> TypeEq for T {
-                    type This = Self;
-                }
-                fn validate_result_error_type<T>()
-                where
-                    T: ?Sized + TypeEq<This = dropshot::HttpError>,
-                {
-                }
-                validate_result_error_type::<
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::E,
-                >();
-            };
-
-            #[allow(non_camel_case_types, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            struct handler_xyz {}
-
-            #[allow(non_upper_case_globals, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            const handler_xyz: handler_xyz = handler_xyz {};
-
-            impl From<handler_xyz>
-                for dropshot::ApiEndpoint<
-                    <RequestContext<std::i32> as dropshot::RequestContextArgument>::Context
-                >
-            {
-                fn from(_: handler_xyz) -> Self {
-                    #[allow(clippy::unused_async)]
-                    async fn handler_xyz(
-                        _rqctx: RequestContext<std::i32>,
-                        q: Query<Q>,
-                    ) ->
-                        Result<HttpResponseOk<()>, HttpError>
-                    {
-                        Ok(())
-                    }
-
-                    const _: fn() = || {
-                        fn future_endpoint_must_be_send<T: ::std::marker::Send>(_t: T) {}
-                        fn check_future_bounds(arg0: RequestContext<std::i32>, arg1: Query<Q>) {
-                            future_endpoint_must_be_send(handler_xyz(arg0, arg1));
-                        }
-                    };
-
-                    dropshot::ApiEndpoint::new(
-                        "handler_xyz".to_string(),
-                        handler_xyz,
-                        dropshot::Method::GET,
-                        "application/json",
-                        "/a/b/c",
-                    )
-                }
-            }
-        };
 
         assert!(errors.is_empty());
-        assert_eq!(expected.to_string(), item.to_string());
+        assert_contents(
+            "tests/output/endpoint_with_query.rs",
+            &prettyplease::unparse(&parse_quote! { #item }),
+        );
     }
 
     #[test]
@@ -757,93 +543,12 @@ mod tests {
             },
         )
         .unwrap();
-        let expected = quote! {
-            const _: fn() = || {
-                struct NeedRequestContext(<RequestContext<()> as dropshot::RequestContextArgument>::Context) ;
-            };
-            const _: fn() = || {
-                fn need_exclusive_extractor<T>()
-                where
-                    T: ?Sized + dropshot::ExclusiveExtractor,
-                {
-                }
-                need_exclusive_extractor::<Query<Q> >();
-            };
-            const _: fn() = || {
-                trait ResultTrait {
-                    type T;
-                    type E;
-                }
-                impl<TT, EE> ResultTrait for Result<TT, EE>
-                where
-                    TT: dropshot::HttpResponse,
-                {
-                    type T = TT;
-                    type E = EE;
-                }
-                struct NeedHttpResponse(
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::T,
-                );
-                trait TypeEq {
-                    type This: ?Sized;
-                }
-                impl<T: ?Sized> TypeEq for T {
-                    type This = Self;
-                }
-                fn validate_result_error_type<T>()
-                where
-                    T: ?Sized + TypeEq<This = dropshot::HttpError>,
-                {
-                }
-                validate_result_error_type::<
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::E,
-                >();
-            };
-
-            #[allow(non_camel_case_types, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            pub(crate) struct handler_xyz {}
-
-            #[allow(non_upper_case_globals, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            pub(crate) const handler_xyz: handler_xyz = handler_xyz {};
-
-            impl From<handler_xyz>
-                for dropshot::ApiEndpoint<
-                    <RequestContext<()> as dropshot::RequestContextArgument>::Context
-                >
-            {
-                fn from(_: handler_xyz) -> Self {
-                    #[allow(clippy::unused_async)]
-                    pub(crate) async fn handler_xyz(
-                        _rqctx: RequestContext<()>,
-                        q: Query<Q>,
-                    ) ->
-                        Result<HttpResponseOk<()>, HttpError>
-                    {
-                        Ok(())
-                    }
-
-                    const _: fn() = || {
-                        fn future_endpoint_must_be_send<T: ::std::marker::Send>(_t: T) {}
-                        fn check_future_bounds(arg0: RequestContext<()>, arg1: Query<Q>) {
-                            future_endpoint_must_be_send(handler_xyz(arg0, arg1));
-                        }
-                    };
-
-                    dropshot::ApiEndpoint::new(
-                        "handler_xyz".to_string(),
-                        handler_xyz,
-                        dropshot::Method::GET,
-                        "application/json",
-                        "/a/b/c",
-                    )
-                }
-            }
-        };
 
         assert!(errors.is_empty());
-        assert_eq!(expected.to_string(), item.to_string());
+        assert_contents(
+            "tests/output/endpoint_pub_crate.rs",
+            &prettyplease::unparse(&parse_quote! { #item }),
+        );
     }
 
     #[test]
@@ -863,84 +568,12 @@ mod tests {
             },
         )
         .unwrap();
-        let expected = quote! {
-            const _: fn() = || {
-                struct NeedRequestContext(<RequestContext<()> as dropshot::RequestContextArgument>::Context) ;
-            };
-            const _: fn() = || {
-                trait ResultTrait {
-                    type T;
-                    type E;
-                }
-                impl<TT, EE> ResultTrait for Result<TT, EE>
-                where
-                    TT: dropshot::HttpResponse,
-                {
-                    type T = TT;
-                    type E = EE;
-                }
-                struct NeedHttpResponse(
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::T,
-                );
-                trait TypeEq {
-                    type This: ?Sized;
-                }
-                impl<T: ?Sized> TypeEq for T {
-                    type This = Self;
-                }
-                fn validate_result_error_type<T>()
-                where
-                    T: ?Sized + TypeEq<This = dropshot::HttpError>,
-                {
-                }
-                validate_result_error_type::<
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::E,
-                >();
-            };
-
-            #[allow(non_camel_case_types, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            struct handler_xyz {}
-
-            #[allow(non_upper_case_globals, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz"]
-            const handler_xyz: handler_xyz = handler_xyz {};
-
-            impl From<handler_xyz>
-                for dropshot::ApiEndpoint<
-                    <RequestContext<()>
-                as dropshot::RequestContextArgument>::Context>
-            {
-                fn from(_: handler_xyz) -> Self {
-                    #[allow(clippy::unused_async)]
-                    async fn handler_xyz(
-                        _rqctx: RequestContext<()>,
-                    ) -> Result<HttpResponseOk<()>, HttpError> {
-                        Ok(())
-                    }
-
-                    const _: fn() = || {
-                        fn future_endpoint_must_be_send<T: ::std::marker::Send>(_t: T) {}
-                        fn check_future_bounds(arg0: RequestContext<()>) {
-                            future_endpoint_must_be_send(handler_xyz(arg0));
-                        }
-                    };
-
-                    dropshot::ApiEndpoint::new(
-                        "handler_xyz".to_string(),
-                        handler_xyz,
-                        dropshot::Method::GET,
-                        "application/json",
-                        "/a/b/c",
-                    )
-                    .tag("stuff")
-                    .tag("things")
-                }
-            }
-        };
 
         assert!(errors.is_empty());
-        assert_eq!(expected.to_string(), item.to_string());
+        assert_contents(
+            "tests/output/endpoint_with_tags.rs",
+            &prettyplease::unparse(&parse_quote! { #item }),
+        );
     }
 
     #[test]
@@ -960,84 +593,12 @@ mod tests {
             },
         )
         .unwrap();
-        let expected = quote! {
-            const _: fn() = || {
-                struct NeedRequestContext(<RequestContext<()> as dropshot::RequestContextArgument>::Context) ;
-            };
-            const _: fn() = || {
-                trait ResultTrait {
-                    type T;
-                    type E;
-                }
-                impl<TT, EE> ResultTrait for Result<TT, EE>
-                where
-                    TT: dropshot::HttpResponse,
-                {
-                    type T = TT;
-                    type E = EE;
-                }
-                struct NeedHttpResponse(
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::T,
-                );
-                trait TypeEq {
-                    type This: ?Sized;
-                }
-                impl<T: ?Sized> TypeEq for T {
-                    type This = Self;
-                }
-                fn validate_result_error_type<T>()
-                where
-                    T: ?Sized + TypeEq<This = dropshot::HttpError>,
-                {
-                }
-                validate_result_error_type::<
-                    <Result<HttpResponseOk<()>, HttpError> as ResultTrait>::E,
-                >();
-            };
-
-            #[allow(non_camel_case_types, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz\nhandle \"xyz\" requests"]
-            struct handler_xyz {}
-
-            #[allow(non_upper_case_globals, missing_docs)]
-            #[doc = "API Endpoint: handler_xyz\nhandle \"xyz\" requests"]
-            const handler_xyz: handler_xyz = handler_xyz {};
-
-            impl From<handler_xyz>
-                for dropshot::ApiEndpoint<
-                    <RequestContext<()>
-                as dropshot::RequestContextArgument>::Context>
-            {
-                fn from(_: handler_xyz) -> Self {
-                    #[allow(clippy::unused_async)]
-                    #[doc = r#" handle "xyz" requests "#]
-                    async fn handler_xyz(
-                        _rqctx: RequestContext<()>,
-                    ) -> Result<HttpResponseOk<()>, HttpError> {
-                        Ok(())
-                    }
-
-                    const _: fn() = || {
-                        fn future_endpoint_must_be_send<T: ::std::marker::Send>(_t: T) {}
-                        fn check_future_bounds(arg0: RequestContext<()>) {
-                            future_endpoint_must_be_send(handler_xyz(arg0));
-                        }
-                    };
-
-                    dropshot::ApiEndpoint::new(
-                        "handler_xyz".to_string(),
-                        handler_xyz,
-                        dropshot::Method::GET,
-                        "application/json",
-                        "/a/b/c",
-                    )
-                    .summary("handle \"xyz\" requests")
-                }
-            }
-        };
 
         assert!(errors.is_empty());
-        assert_eq!(expected.to_string(), item.to_string());
+        assert_contents(
+            "tests/output/endpoint_with_doc.rs",
+            &prettyplease::unparse(&parse_quote! { #item }),
+        );
     }
 
     #[test]
